+++ neo/framework/CmdSystem.cpp
@@ -527,7 +527,7 @@ void idCmdSystemLocal::InsertCommandText( const char *text ) {
 	int		len;
 	int		i;
 
-	len = strlen( text ) + 1;
+	len = (int)strlen( text ) + 1;
 	if ( len + textLength > (int)sizeof( textBuf ) ) {
 		common->Printf( "idCmdSystemLocal::InsertText: buffer overflow\n" );
 		return;
@@ -557,7 +557,7 @@ Adds command text at the end of the buffer, does NOT add a final \n
 void idCmdSystemLocal::AppendCommandText( const char *text ) {
 	int l;
 	
-	l = strlen( text );
+	l = (int)strlen( text );
 
 	if ( textLength + l >= (int)sizeof( textBuf ) ) {
 		common->Printf( "idCmdSystemLocal::AppendText: buffer overflow\n" );
